backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
Werkzeug==3.0.1
gunicorn==21.2.0

